added new files
Added madical.html aswell as .js, same for building, finance and education. updatet index.html a bit, changed name on main frontend for calulators to calculators.html to make it better looking. and yeah made som changes and added accounts to the system/server/website or what ever and yea